rename: apply kebab case
useAgreementStore.tsx -> use-agreement-store.tsx
